refactor: split PD disaggregation router into focused files
Extract scoring types and HTTP prefill logic from the monolithic
pd_disaggregation.go into a dedicated sub-package and a separate file:

- algorithms/pd/prefill_scorer.go: PrefillScorePolicy / PrefillScorer
  interfaces with prefix_cache and least_request implementations
- algorithms/pd/trackers.go: PrefillRequestTracker and PendingDecodeTracker
- algorithms/pd/decode_scorer.go: updated package-level godoc (authoritative)
- algorithms/pd_prefill_request.go: six *pdRouter methods for HTTP prefill
  execution, payload preparation, and KV/disagg param injection

Update test and benchmark files to use exported types from the pd package.
Add comprehensive godoc across all four affected files.
